+++ engine/apps/public_api/serializers/personal_notification_rules.py
@@ -5,23 +5,21 @@ from rest_framework import exceptions, serializers
 
 from apps.base.models import UserNotificationPolicy
 from apps.base.models.user_notification_policy import NotificationChannelPublicAPIOptions
+from apps.public_api.serializers.ordered_model import OrderedModelSerializer
 from common.api_helpers.custom_fields import UserIdField
 from common.api_helpers.exceptions import BadRequest
 from common.api_helpers.mixins import EagerLoadingMixin
 
 
-class PersonalNotificationRuleSerializer(EagerLoadingMixin, serializers.ModelSerializer):
+class PersonalNotificationRuleSerializer(EagerLoadingMixin, OrderedModelSerializer):
     id = serializers.CharField(read_only=True, source="public_primary_key")
     user_id = UserIdField(required=True, source="user")
-    position = serializers.IntegerField(required=False, source="order")
     type = serializers.CharField(
         required=False,
     )
-
     duration = serializers.ChoiceField(
         required=False, source="wait_delay", choices=UserNotificationPolicy.DURATION_CHOICES
     )
-    manual_order = serializers.BooleanField(default=False, write_only=True)
 
     SELECT_RELATED = ["user"]
 
@@ -31,7 +29,8 @@ class PersonalNotificationRuleSerializer(EagerLoadingMixin, serializers.ModelSer
 
     class Meta:
         model = UserNotificationPolicy
-        fields = ["id", "user_id", "position", "type", "duration", "manual_order", "important"]
+        fields = ["id", "user_id", "type", "duration", "important"]
+        fields += ["position", "manual_order"]  # fields from OrderedModelSerializer
 
     def create(self, validated_data):
         if "type" not in validated_data:
@@ -44,17 +43,7 @@ class PersonalNotificationRuleSerializer(EagerLoadingMixin, serializers.ModelSer
         if "wait_delay" in validated_data and validated_data["step"] != UserNotificationPolicy.Step.WAIT:
             raise exceptions.ValidationError({"duration": "Duration can't be set"})
 
-        # Remove "manual_order" and "order" fields from validated_data, so they are not passed to create method.
-        # Policies are always created at the end of the list, and then moved to the desired position by _adjust_order.
-        manual_order = validated_data.pop("manual_order")
-        order = validated_data.pop("order", None)
-
-        instance = UserNotificationPolicy.objects.create(**validated_data)
-
-        if order is not None:
-            self._adjust_order(instance, manual_order, order, created=True)
-
-        return instance
+        return super().create(validated_data)
 
     def to_internal_value(self, data):
         if "duration" in data:
@@ -119,33 +108,6 @@ class PersonalNotificationRuleSerializer(EagerLoadingMixin, serializers.ModelSer
 
         raise exceptions.ValidationError({"type": "Invalid type"})
 
-    @staticmethod
-    def _adjust_order(instance, manual_order, order, created):
-        # Passing order=-1 means that the policy should be moved to the end of the list.
-        if order == -1:
-            if created:
-                # The policy was just created, so it is already at the end of the list.
-                return
-
-            order = instance.max_order()
-            # max_order() can't be None here because at least one instance exists – the one we are moving.
-            assert order is not None
-
-        # Negative order is not allowed.
-        if order < 0:
-            raise BadRequest(detail="Invalid value for position field")
-
-        # manual_order=True is intended for use by Terraform provider only, and is not a documented feature.
-        # Orders are swapped instead of moved when using Terraform, because Terraform may issue concurrent requests
-        # to create / update / delete multiple policies. "Move to" operation is not deterministic in this case, and
-        # final order of policies may be different depending on the order in which requests are processed. On the other
-        # hand, the result of concurrent "swap" operations is deterministic and does not depend on the order in
-        # which requests are processed.
-        if manual_order:
-            instance.swap(order)
-        else:
-            instance.to(order)
-
 
 class PersonalNotificationRuleUpdateSerializer(PersonalNotificationRuleSerializer):
     user_id = UserIdField(read_only=True, source="user")
@@ -164,11 +126,5 @@ class PersonalNotificationRuleUpdateSerializer(PersonalNotificationRuleSerialize
         else:
             if "wait_delay" in validated_data and instance.step != UserNotificationPolicy.Step.WAIT:
                 raise exceptions.ValidationError({"duration": "Duration can't be set"})
-
-        # Remove "manual_order" and "order" fields from validated_data, so they are not passed to update method.
-        manual_order = validated_data.pop("manual_order")
-        order = validated_data.pop("order", None)
-        if order is not None:
-            self._adjust_order(instance, manual_order, order, created=False)
 
         return super().update(instance, validated_data)
